requirements.txt
Pillow==8.0.1 
plyer==2.0.0